+++ apps/video/app/demo.tsx
@@ -31,6 +31,9 @@ export const Demo = () => {
         poster="https://images.unsplash.com/photo-1484291470158-b8f8d608850d?w=900&auto=format&fit=crop&q=60&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxzZWFyY2h8OHx8b2NlYW58ZW58MHx8MHx8fDA%3D"
         ratio="16/9"
         // autoPlay="force"
+        config={{
+          range: [10, 20],
+        }}
       >
         {({
           isPlaying,