+++ packages/video/src/hooks/use-range.tsx
@@ -1,0 +1,3 @@
+export const useRange = () => {
+  return <div>useRange</div>;
+};