+++ apps/video/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes